// src/api/models/DsChat.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Mailing app
 * my api
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DsChat
 */
export interface DsChat {
    /**
     * 
     * @type {number}
     * @memberof DsChat
     */
    friends?: number;
    /**
     * 
     * @type {number}
     * @memberof DsChat
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof DsChat
     */
    img?: string;
    /**
     * 
     * @type {string}
     * @memberof DsChat
     */
    info?: string;
    /**
     * 
     * @type {boolean}
     * @memberof DsChat
     */
    isDelete?: boolean;
    /**
     * 
     * @type {string}
     * @memberof DsChat
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof DsChat
     */
    nickname?: string;
    /**
     * 
     * @type {number}
     * @memberof DsChat
     */
    subscribers?: number;
}

/**
 * Check if a given object implements the DsChat interface.
 */
export function instanceOfDsChat(value: object): value is DsChat {
    return true;
}

export function DsChatFromJSON(json: any): DsChat {
    return DsChatFromJSONTyped(json, false);
}

export function DsChatFromJSONTyped(json: any, ignoreDiscriminator: boolean): DsChat {
    if (json == null) {
        return json;
    }
    return {
        
        'friends': json['friends'] == null ? undefined : json['friends'],
        'id': json['id'] == null ? undefined : json['id'],
        'img': json['img'] == null ? undefined : json['img'],
        'info': json['info'] == null ? undefined : json['info'],
        'isDelete': json['isDelete'] == null ? undefined : json['isDelete'],
        'name': json['name'] == null ? undefined : json['name'],
        'nickname': json['nickname'] == null ? undefined : json['nickname'],
        'subscribers': json['subscribers'] == null ? undefined : json['subscribers'],
    };
}

export function DsChatToJSON(json: any): DsChat {
    return DsChatToJSONTyped(json, false);
}

export function DsChatToJSONTyped(value?: DsChat | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'friends': value['friends'],
        'id': value['id'],
        'img': value['img'],
        'info': value['info'],
        'isDelete': value['isDelete'],
        'name': value['name'],
        'nickname': value['nickname'],
        'subscribers': value['subscribers'],
    };
}