// src/api/models/DsChatRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Mailing app
 * my api
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DsChatRequest
 */
export interface DsChatRequest {
    /**
     * 
     * @type {number}
     * @memberof DsChatRequest
     */
    friends?: number;
    /**
     * 
     * @type {string}
     * @memberof DsChatRequest
     */
    info?: string;
    /**
     * 
     * @type {string}
     * @memberof DsChatRequest
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof DsChatRequest
     */
    nickname?: string;
    /**
     * 
     * @type {number}
     * @memberof DsChatRequest
     */
    subscribers?: number;
}

/**
 * Check if a given object implements the DsChatRequest interface.
 */
export function instanceOfDsChatRequest(value: object): value is DsChatRequest {
    return true;
}

export function DsChatRequestFromJSON(json: any): DsChatRequest {
    return DsChatRequestFromJSONTyped(json, false);
}

export function DsChatRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): DsChatRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'friends': json['friends'] == null ? undefined : json['friends'],
        'info': json['info'] == null ? undefined : json['info'],
        'name': json['name'] == null ? undefined : json['name'],
        'nickname': json['nickname'] == null ? undefined : json['nickname'],
        'subscribers': json['subscribers'] == null ? undefined : json['subscribers'],
    };
}

export function DsChatRequestToJSON(json: any): DsChatRequest {
    return DsChatRequestToJSONTyped(json, false);
}

export function DsChatRequestToJSONTyped(value?: DsChatRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'friends': value['friends'],
        'info': value['info'],
        'name': value['name'],
        'nickname': value['nickname'],
        'subscribers': value['subscribers'],
    };
}